Add files via upload
Checks and flags APIs whose response time is slow. It gives a final count on a notepad printout. This check had two branches from the API. One branch had and ID and one branch did not have an ID but the response times were still checked for each one. If the API fails, the OUT OF SPECS is in all caps to be easier to find.